34 - Refactor graphics controllers to use protocol type Replaced XRGeometryController parameter types with id<GraphicGeometrySource> in all graphics classes and initializers. This change increases flexibility by allowing any object conforming to the GraphicGeometrySource protocol to be used as a controller, improving code modularity and testability.

+++ PaleoRose/Classes/Graphics/GraphicGeometrySource.h
@@ -31,7 +31,7 @@
 - (NSRect)circleRectForGeometryPercent:(float)percent;
 - (NSPoint)rotationOfPoint:(NSPoint)thePoint byAngle:(double)angle;
 - (double)degreesFromRadians:(double)radians;
-- (double)radiusOfRelativePercent:(float)percent;
-- (CGFloat)unrestrictedRadiusOfRelativePercent:(CGFloat)percent;
+- (double)radiusOfRelativePercent:(double)percent;
+- (CGFloat)unrestrictedRadiusOfRelativePercent:(double)percent;
 
 @end
